requirements.txt
streamlit
openai
python-dotenv
streamlit
speechrecognition
pyaudio